Initial commit
Basic, cross-Browser and cross-Platform functionality
Yet still full of bugs. I.e. in html-mode the mouseout event get's
triggered when entering a target, not leaving it. The Cursor won't get
inherited. When leaving the window, the cursor stays on the document.
And: Not yet tested in any version of Internet Explorer.
So, still a lot to do.